Add tests for config pruning and defaulting
